Feat/homegeofence map

+++ TeslaSolarCharger/Shared/Contracts/IConfigurationWrapper.cs
@@ -99,8 +99,8 @@ public interface IConfigurationWrapper
     int? MaxInverterAcPower();
     string? BleBaseUrl();
     bool SendTeslaApiStatsToBackend();
-    decimal HomeGeofenceLongitude();
-    decimal HomeGeofenceLatitude();
+    double HomeGeofenceLongitude();
+    double HomeGeofenceLatitude();
     int HomeGeofenceRadius();
     bool ShouldUseFakeSolarValues();
     int MaxTravelSpeedMetersPerSecond();
